Fix: Applying a patch in Windows throws PatchApplyException
The exception is caused because the line endings differ between patch (no matter if it's created by git or JGit, both have \n) and working tree (normally \r\n on windows).
The fix does a trim before the comparison (so the lines are still considered the same if they only differ in trailing/leading whitespaces) and, if the OS is Windows and the \r is missing, appends the \r in the patched lines.